refactor: Moved menu from Main to its own class
